AAP: Exclude tests which use -f[no-]delete-null-pointer-checks
This adds a new annotation to check whether the
-f[no-]delete-null-pointer-checks flag is supported by the compiler.
Various tests have been updated to use this annotation.

Although -fdelete-null-pointer-checks is a gcc-exclusive flag,
it is not unlikely that it will eventually be supported in
clang and other compilers as it is important to be able to
disable the optimization when building the linux kernel. For
this reason, a new check has been added, instead of just using
the check for gcc internals.

gcc/testsuite/ChangeLog.Embecosm:

	* c-c++-common/pr27336.c: Mark test with
	dg-require-effective-target annotation to check for support
	for the -fdelete-null-pointer-checks optimization.
	* gcc.dg/pr22335-1.c: Likewise.
	* gcc.dg/pr38984.c: Likewise.
	* gcc.dg/pr42084.c: Likewise.
	* gcc.dg/pr43971.c: Likewise.
	* gcc.dg/pr44024.c: Likewise.
	* gcc.dg/pr49235.c: Likewise.
	* gcc.dg/torture/pr51071-2.c: Likewise.
	* lib/target-supports.exp
	(check_effective_target_delete_null_pointer_checks): Check
	whether the compiler supports the -fdelete-null-pointer-checks
	optimization.