Fix possible div-by-zero in gcc.c-torture/execute/pr53645.c
The test declares an int vector of 16 bytes. On architectures
where 'int' is 2 bytes this corresponds to an 8 element
vector, however in various places vectors are explicitly
initialized with only 4 elements. As a consequence the remaining
elements are initialized to zero. These zero-initialized
elements are then used on the right hand side of a divide,
which is undefined.

gcc/testsuite/ChangeLog:

	* gcc.c-torture/execute/pr53645.c (UV SV): Ensure
	that typdef'd vector types are only 4 elements wide, to
	match the size of the statically initialized vectors
	in the test.