Annotate bitfield tests as requiring a gcc frontend
The behavior of bitfields of implementation defined type differs
between GCC and Clang. GCC performs the operation at the precision
of the bitfield width in bits, whereas Clang performs the
operation at the precision of the underlying type.

See https://gcc.gnu.org/ml/gcc/2017-10/msg00191.html for a
discussion of the difference.

gcc/testsuite/ChangeLog:

	* gcc.c-torture/execute/bitfld-3.c: Annotate as only being
	supported with a gcc frontend.
	* gcc.c-torture/execute/bitfld-5.c: Likewise.
	* gcc.c-torture/execute/pr32244-1.c: Likewise.
	* gcc.c-torture/execute/pr34971.c: Likewise.