Ensure -fpeel-loops and -ftracer are not used for Clang
To reduce the amount of failures caused by unrecognized optimization
arguments on Clang, do not pass -fpeel-loops and -ftracer when the
target is an llvm compiler. The check_effective_target_llvm is used
here, however ideally this check will be removed, so this patch should
be changed if/when this happens.

gcc/testsuite/Changelog.Embecosm:

	* lib/gcc-dg.exp: Don't pass -fpeel-loops and -ftracer when
	using clang.
	* lib/c-torture.exp: Likewise.